requirements.txt
fastapi
uvicorn[standard]
pymongo
motor
mysql-connector-python
sqlalchemy
pymysql
redis
aioredis
pyjwt